refactor: organize distribution components code
