+++ src/components/templates/DistributionProducts.tsx
@@ -1,4 +1,3 @@
-import { StaticImageData } from 'next/image';
 import { useTranslation } from 'next-i18next';
 import * as React from 'react';
 
@@ -8,21 +7,14 @@ import { Body, Title, UnderlineLink } from '@/components/ui';
 
 import { company } from '@/constant/company';
 
+import { Product } from '@/types/Product';
+
 import ProductBissinger from '/public/images/distribution/products/distribution_products-bissinger.webp';
 import ProductEberle from '/public/images/distribution/products/distribution_products-eberle.webp';
 import ProductHupfer from '/public/images/distribution/products/distribution_products-hupfer.webp';
 import ProductMedicon from '/public/images/distribution/products/distribution_products-medicon.webp';
 import ProductNouvag from '/public/images/distribution/products/distribution_products-nouvag.webp';
 import ProductTontarra from '/public/images/distribution/products/distribution_products-tontarra.webp';
-
-export interface Product {
-  title: string;
-  text: string;
-  image: StaticImageData;
-  manufacturer: string;
-  url: string;
-  manual: string;
-}
 
 const products: Product[] = [
   {
@@ -78,55 +70,68 @@ const products: Product[] = [
 export default function DistributionProducts() {
   const { t } = useTranslation('distribution');
 
+  const ProductsHelp = () => {
+    return (
+      <>
+        <Title size='one' renderAs='h3'>
+          {t('content.products.title')}
+        </Title>
+        <Body size='sm' color='light'>
+          {t('content.products.text')}
+        </Body>
+        <UnderlineLink underline='always' href='/kontakt'>
+          {t('content.products.button')}
+        </UnderlineLink>
+        <div className='mt-8 hidden bg-gray-100 p-3 xl:block'>
+          <Body>
+            {company.street} <br />
+            {company.city}
+          </Body>
+          <Body>
+            <UnderlineLink href={`tel:${company.phone}`} underline='hover'>
+              {company.phone}
+            </UnderlineLink>
+            <br />
+
+            <UnderlineLink href={'mailto:' + company.email} underline='hover'>
+              {company.email}
+            </UnderlineLink>
+          </Body>
+        </div>
+      </>
+    );
+  };
+
+  const ProductGrid = () => {
+    return (
+      <>
+        {products.map((product) => (
+          <ProductCard
+            manufacturer={product.manufacturer}
+            key={product.manufacturer}
+            // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+            // @ts-expect-error
+            title={t(product.title)}
+            // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+            // @ts-expect-error
+            description={t(product.text)}
+            image={product.image}
+            url={product.url}
+          />
+        ))}
+      </>
+    );
+  };
+
   return (
     <section className='pt-16 md:pt-24 lg:pt-32'>
       <Container>
         <div className='flex flex-col gap-10 bg-white xl:flex-row'>
           <div className='mx-auto xl:max-w-[480px]'>
-            <Title size='one' renderAs='h3'>
-              {t('content.products.title')}
-            </Title>
-            <Body size='sm' color='light'>
-              {t('content.products.text')}
-            </Body>
-            <UnderlineLink underline='always' href='/kontakt'>
-              {t('content.products.button')}
-            </UnderlineLink>
-            <div className='mt-8 hidden bg-gray-100 p-3 xl:block'>
-              <Body>
-                {company.street} <br />
-                {company.city}
-              </Body>
-              <Body>
-                <UnderlineLink href={`tel:${company.phone}`} underline='hover'>
-                  {company.phone}
-                </UnderlineLink>
-                <br />
-
-                <UnderlineLink
-                  href={'mailto:' + company.email}
-                  underline='hover'
-                >
-                  {company.email}
-                </UnderlineLink>
-              </Body>
-            </div>
+            <ProductsHelp />
           </div>
           <div className='mx-auto grid gap-4 md:grid-cols-2'>
-            {products.map((product) => (
-              <ProductCard
-                manufacturer={product.manufacturer}
-                key={product.manufacturer}
-                // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-                // @ts-expect-error
-                title={t(product.title)}
-                // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-                // @ts-expect-error
-                description={t(product.text)}
-                image={product.image}
-                url={product.url}
-              />
-            ))}
+            <ProductGrid />
           </div>
         </div>
       </Container>
